backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.12.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Django==5.2.3
django-cors-headers==4.3.1
django-ninja==1.4.3
et_xmlfile==2.0.0
frozenlist==1.7.0
fuzzywuzzy==0.18.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
lxml==4.9.3
multidict==6.6.4
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.30.0
selenium==4.23.1
amazoncaptcha==0.5.11
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
whitenoise==6.9.0
yarl==1.20.1
google-api-python-client==2.137.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
gspread==6.1.4
httplib2==0.22.0
